backend/backend_2/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
alembic  # optionnel pour migrations
faker
PyJWT

